backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
deepface==0.0.79
tensorflow==2.17.1
numpy==1.24.3
pillow==10.1.0